+++ include/H2DE/H2DE/H2DE_timeline.h
@@ -1,7 +1,7 @@
 #ifndef H2DE_TIMELINE__H
 #define H2DE_TIMELINE__H
 
-#include <H2DE/H2DE_engine.h>
+#include <H2DE/H2DE/H2DE_engine.h>
 
 #undef max