+++ src/objects/H2DE_button_object.cpp
@@ -1,4 +1,4 @@
-#include "H2DE/objects/H2DE_button_object.h"
+#include "H2DE/H2DE/objects/H2DE_button_object.h"
 
 // INIT
 H2DE_ButtonObject::H2DE_ButtonObject(H2DE_Engine* engine, H2DE_ObjectData od, H2DE_ButtonObjectData b) : H2DE_Object(engine, od), bod(b) {
@@ -20,68 +20,31 @@ H2DE_ButtonObject::~H2DE_ButtonObject() {
 
 // UPDATE
 void H2DE_ButtonObject::update() {
-    // if (!bod.onMouseDown&& !bod.onMouseUp && !bod.onHover && !bod.onBlur) {
-    //     return;
-    // }
 
-    // H2DE_LevelPos mousePos = H2DE_GetMousePos(engine, od.absolute);
-    // H2DE_LevelPos pos = od.pos;
-    // bool clicked = false;
-    // bool hovered = false;
-
-    // for (auto [name, hitbox] : od.hitboxes) {
-    //     if (hitbox.rect.addPos(pos).collides(mousePos)) {
-    //         if (bod.onclick && !clicked && engine->click.has_value()) {
-    //             bod.onclick();
-    //             clicked = true;
-    //         }
-
-    //         if (bod.onhover && !hover && !hovered) {
-    //             hover = true;
-    //             hovered = true;
-    //             bod.onhover();
-    //         }
-    //     }
-    // }
-
-    // if (!hovered && hover && bod.onout) {
-    //     int nbOut = 0;
-
-    //     for (auto [name, hitbox] : od.hitboxes) {
-    //         if (!hitbox.rect.addPos(pos).collides(mousePos)) {
-    //             nbOut++;
-    //         }
-    //     }
-
-    //     if (nbOut >= od.hitboxes.size()) {
-    //         bod.onout();
-    //         hover = false;
-    //     }
-    // }
 }
 
 // CALLS
 void H2DE_ButtonMouseDown(H2DE_ButtonObject* button) {
     if (button->bod.onMouseDown) {
-        button->bod.onMouseDown();
+        button->bod.onMouseDown(button);
     }
 }
 
 void H2DE_ButtonMouseUp(H2DE_ButtonObject* button) {
     if (button->bod.onMouseUp) {
-        button->bod.onMouseUp();
+        button->bod.onMouseUp(button);
     }
 }
 
 void H2DE_ButtonHover(H2DE_ButtonObject* button) {
     if (button->bod.onHover) {
-        button->bod.onHover();
+        button->bod.onHover(button);
     }
 }
 
 void H2DE_ButtonBlur(H2DE_ButtonObject* button) {
     if (button->bod.onBlur) {
-        button->bod.onBlur();
+        button->bod.onBlur(button);
     }
 }
 
@@ -104,18 +67,18 @@ H2DE_Surface* H2DE_GetButtonSurface(const H2DE_ButtonObject* button) {
 }
 
 // SETTER
-void H2DE_SetButtonOnMouseDown(H2DE_ButtonObject* button, const std::function<void()>& onMouseDown) {
+void H2DE_SetButtonOnMouseDown(H2DE_ButtonObject* button, const std::function<void(H2DE_Object*)>& onMouseDown) {
     button->bod.onMouseDown = onMouseDown;    
 }
 
-void H2DE_SetButtonOnMouseUp(H2DE_ButtonObject* button, const std::function<void()>& onMouseUp) {
+void H2DE_SetButtonOnMouseUp(H2DE_ButtonObject* button, const std::function<void(H2DE_Object*)>& onMouseUp) {
     button->bod.onMouseUp = onMouseUp;
 }
 
-void H2DE_SetButtonOnHover(H2DE_ButtonObject* button, const std::function<void()>& onHover) {
+void H2DE_SetButtonOnHover(H2DE_ButtonObject* button, const std::function<void(H2DE_Object*)>& onHover) {
     button->bod.onHover = onHover;
 }
 
-void H2DE_SetButtonOnBlur(H2DE_ButtonObject* button, const std::function<void()>& onBlur) {
+void H2DE_SetButtonOnBlur(H2DE_ButtonObject* button, const std::function<void(H2DE_Object*)>& onBlur) {
     button->bod.onBlur = onBlur;
 }