+++ src/surface/H2DE_surface.cpp
@@ -1,4 +1,4 @@
-#include "H2DE/surfaces/H2DE_surface.h"
+#include "H2DE/H2DE/surfaces/H2DE_surface.h"
 
 // INIT
 H2DE_Surface::H2DE_Surface(H2DE_Engine* e, const H2DE_SurfaceData& s) : engine(e), sd(s) {