+++ src/utils/H2DE_rect.cpp
@@ -1,4 +1,4 @@
-#include "H2DE/H2DE_utils.h"
+#include "H2DE/H2DE/H2DE_utils.h"
 
 template struct H2DE_Rect<int>;
 template struct H2DE_Rect<float>;